+++ config/routes/data.rb
@@ -300,6 +300,7 @@ resources :dwc_occurrences, only: [:create] do
     get 'collector_id_metadata', defaults: {format: :json}
     get 'download'
     post 'sweep', as: 'sweep_stale' # TODO: ultimately should not be required
+    get :attributes, defaults: {format: :json}
   end
 end