+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 - Reasonable min/max elevations hard validations 
 - Increased scope of string cleaning [#3947]
+- DwcOccurrence filter on all attributes
 
 ### Fixed